[feat]: Optimized code and updated readme
