+++ readme/img/claude_opening_setting.png
